requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
BaseModel==20220524.542
bcrypt==3.2.0
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach==5.0.0
boto3==1.21.39
botocore==1.24.39
Bottleneck==1.3.4
brotlipy==0.7.0
cachetools==5.1.0
catboost==1.0.5
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
click==8.1.2
conda==4.9.2
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.11.0
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.1.0
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fastapi==0.78.0
fastjsonschema==2.15.3
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
fonttools==4.33.2
gevent==21.12.0
gitdb==4.0.9
GitPython==3.1.27
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
graphviz==0.20
greenlet==1.1.2
h11==0.13.0
huggingface-hub==0.5.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-resources==5.7.0
inotify-simple==1.2.1
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.7.0
jedi @ file:///tmp/build/80754af9/jedi_1606932564285/work
Jinja2 @ file:///home/linux1/recipes/ci/jinja2_1610990516718/work
jmespath==1.0.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyter-server==1.16.0
jupyterlab==3.3.3
jupyterlab-pygments==0.2.1
jupyterlab-server==2.12.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
lightgbm==3.3.2
MarkupSafe==1.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclassic==0.3.7
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
notebook-shim==0.1.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
olefile==0.46
packaging==21.3
pandas==1.0.1
pandocfilters==1.5.0
paramiko==2.10.3
parmap==1.5.3
parso==0.7.0
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1609786786540/work
pkginfo==1.7.0
plotly==5.8.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.20.0
psutil @ file:///tmp/build/80754af9/psutil_1598370257551/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.9.1
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
pymongo==4.1.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-etcd==0.4.5
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyYAML==5.3.1
pyzmq==22.3.0
regex==2022.3.15
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
retrying==1.3.3
rsa==4.8
ruamel-yaml==0.15.87
s3transfer==0.5.2
sacremoses==0.0.49
sagemaker-training==4.1.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.10
setproctitle==1.2.3
shortuuid==1.0.8
six @ file:///tmp/build/80754af9/six_1605205327372/work
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1607965878077/work
SQLAlchemy==1.4.36
stack-data==0.2.0
starlette==0.19.1
tenacity==8.0.1
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
torch==1.7.1
torchelastic==0.2.1
torchvision==0.8.2
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets==5.1.1
transformers==4.18.0
typing-extensions==4.2.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
uvicorn==0.17.6
wandb==0.12.15
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
widgetsnbextension==3.6.0
xgboost==1.6.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
